requirements.txt
click>=8.0.0
PyMuPDF>=1.20.0
pillow>=9.0.0
pytesseract>=0.3.10
pandas>=1.5.0
tabula-py>=2.5.0
camelot-py[cv]>=0.10.0
tabulate>=0.8.10 